rename lodmanager to lod
