+++ examples/room/lod/lod/manager.go
