+++ examples/room/lod/lod/manager_test.go
@@ -1,4 +1,4 @@
-package lodmanager
+package lod
 
 import (
 	"testing"
