+++ examples/room/lod/lod/metrics.go
@@ -1,4 +1,4 @@
-package lodmanager
+package lod
 
 import (
 	"github.com/Diarkis/diarkis"
